Add link-preview metadata to the poll detail route
Path-form poll URLs (/g/<group>/p/<poll>) had no generateMetadata, so
sharing one to iMessage/Slack showed a generic title instead of the
poll's. Split the route into a server-component shell (page.tsx) that
fetches the public /preview endpoint with ?p=<pollShortId> and surfaces
the poll title, plus the existing client UI moved to PollDetailPage.tsx.
Mirrors the group route's shell split.
